feat: fix archives names
